Renamed users app to researchers
